+++ src/plugins/JavaScriptPlugin.ts
@@ -2,7 +2,7 @@
  *  gbm Plugin - JavaScript
  */
 
-import {Options, Slot, Stream} from "../core/types";
+import {BuildConfig, GulpStream, Options, Slot} from "../core/types";
 import {GBuilder} from "../core/builder";
 import {GPlugin} from "../core/plugin";
 import {pick} from "../core/utils";
@@ -10,18 +10,18 @@ import {pick} from "../core/utils";
 export class JavaScriptPlugin extends GPlugin {
   constructor(options:Options={}, slots: Slot|Slot[]='build') { super(options, slots); }
 
-  process(stream:Stream, mopts:Options, conf:Options, slot:Slot, builder:GBuilder) {
-    let opts = Object.assign({}, pick(conf.buildOptions, 'lint', 'rename'), this.options);
+  OnStream(stream:GulpStream, mopts:Options, conf:BuildConfig, slot:Slot, builder:GBuilder) {
+    let opts = Object.assign({}, pick(conf.buildOptions || {}, 'lint', 'rename'), this.options);
 
     // check lint option
     if (opts.lint) {
       const esLint = require('gulp-eslint');
       let lintExtra = mopts.eslintExtra || {};
-      stream = stream && stream.pipe(esLint(mopts.eslint))
+      stream = stream.pipe(esLint(mopts.eslint))
         .pipe(esLint.format(lintExtra.format))
         .pipe(esLint.failAfterError());
     }
-    return stream && stream.pipe(require('gulp-babel')(mopts.babel));
+    return stream.pipe(require('gulp-babel')(mopts.babel));
   }
 }